+++ HISTORY.md
@@ -3,6 +3,7 @@ unreleased
 
   * Drop support for Node.js 0.6
   * Remove `status[code]` exports; use `status.message[code]`
+  * Remove `status[msg]` exports; use `status.code[msg]`
   * Rename `425 Unordered Collection` to standard `425 Too Early`
   * Rename `STATUS_CODES` export to `message`